chore: relocate spw vendor
